adk/requirements.txt
google-adk[extensions]
markdown
litellm